# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Service {
  slug: string;
  name: string;
  price: string;
  description: string;
  features: string[];
}

export interface VehicleInput {
  /** @minLength 1 */
  make: string;
  /** @minLength 1 */
  model: string;
  /**
     * @minimum 1900
     * @maximum 2100
     */
  year: number;
  engine: string;
  /** @minimum 0 */
  mileage: number;
  fuel: string;
  transmission: string;
  registration: string;
  /** @nullable */
  vin?: string | null;
}

export type ContactInputPreferredContact = typeof ContactInputPreferredContact[keyof typeof ContactInputPreferredContact];


export const ContactInputPreferredContact = {
  phone: 'phone',
  whatsapp: 'whatsapp',
  email: 'email',
} as const;

export interface ContactInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 5 */
  phone: string;
  /** Email address */
  email: string;
  preferredContact: ContactInputPreferredContact;
}

export type MediaInputKind = typeof MediaInputKind[keyof typeof MediaInputKind];


export const MediaInputKind = {
  photo: 'photo',
  video: 'video',
} as const;

export interface MediaInput {
  name: string;
  contentType: string;
  /** @minimum 1 */
  size: number;
  kind: MediaInputKind;
}

export type CaseInputDrivingStatus = typeof CaseInputDrivingStatus[keyof typeof CaseInputDrivingStatus];


export const CaseInputDrivingStatus = {
  yes: 'yes',
  unsure: 'unsure',
  no: 'no',
} as const;

export interface CaseInput {
  vehicle: VehicleInput;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  symptoms: string;
  faultCodes: string[];
  media: MediaInput[];
  drivingStatus: CaseInputDrivingStatus;
  contact: ContactInput;
  consentAccepted: true;
}

export type AssessmentUrgency = typeof AssessmentUrgency[keyof typeof AssessmentUrgency];


export const AssessmentUrgency = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Assessment {
  urgency: AssessmentUrgency;
  possibleCauses: string[];
  recommendedSteps: string[];
  customerExplanation: string;
  disclaimer: string;
}

export type CaseResultStatus = typeof CaseResultStatus[keyof typeof CaseResultStatus];


export const CaseResultStatus = {
  NEW: 'NEW',
  PRELIMINARY_REVIEW: 'PRELIMINARY_REVIEW',
  DIAGNOSIS_REQUESTED: 'DIAGNOSIS_REQUESTED',
} as const;

export type CaseResultDrivingStatus = typeof CaseResultDrivingStatus[keyof typeof CaseResultDrivingStatus];


export const CaseResultDrivingStatus = {
  yes: 'yes',
  unsure: 'unsure',
  no: 'no',
} as const;

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  REQUESTED: 'REQUESTED',
  CONFIRMED: 'CONFIRMED',
} as const;

export interface Appointment {
  id: string;
  caseId: string;
  service: string;
  preferredDate: string;
  preferredTime: string;
  address: string;
  note: string;
  status: AppointmentStatus;
}

export interface CaseResult {
  id: string;
  caseNumber: string;
  status: CaseResultStatus;
  createdAt: string;
  vehicle: VehicleInput;
  category: string;
  symptoms: string;
  faultCodes: string[];
  media: MediaInput[];
  drivingStatus: CaseResultDrivingStatus;
  contact: ContactInput;
  assessment: Assessment;
  appointment: Appointment | null;
}

export interface AppointmentInput {
  preferredDate: string;
  preferredTime: string;
  /** @minLength 1 */
  address: string;
  note?: string;
}