# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from './appointmentStatus';

export interface Appointment {
  id: string;
  caseId: string;
  service: string;
  preferredDate: Date;
  preferredTime: string;
  address: string;
  note: string;
  status: AppointmentStatus;
}