# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  preferredDate: Date;
  preferredTime: string;
  /** @minLength 1 */
  address: string;
  note?: string;
}