# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseInputDrivingStatus } from './caseInputDrivingStatus';
import type { ContactInput } from './contactInput';
import type { MediaInput } from './mediaInput';
import type { VehicleInput } from './vehicleInput';

export interface CaseInput {
  vehicle: VehicleInput;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  symptoms: string;
  faultCodes: string[];
  media: MediaInput[];
  drivingStatus: CaseInputDrivingStatus;
  contact: ContactInput;
  consentAccepted: true;
}