# lib/api-zod/src/generated/types/caseResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';
import type { Assessment } from './assessment';
import type { CaseResultDrivingStatus } from './caseResultDrivingStatus';
import type { CaseResultStatus } from './caseResultStatus';
import type { ContactInput } from './contactInput';
import type { MediaInput } from './mediaInput';
import type { VehicleInput } from './vehicleInput';

export interface CaseResult {
  id: string;
  caseNumber: string;
  status: CaseResultStatus;
  createdAt: Date;
  vehicle: VehicleInput;
  category: string;
  symptoms: string;
  faultCodes: string[];
  media: MediaInput[];
  drivingStatus: CaseResultDrivingStatus;
  contact: ContactInput;
  assessment: Assessment;
  appointment: Appointment | null;
}