# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInputPreferredContact } from './contactInputPreferredContact';

export interface ContactInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 5 */
  phone: string;
  /** Email address */
  email: string;
  preferredContact: ContactInputPreferredContact;
}