# lib/api-zod/src/generated/types/mediaInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaInputKind } from './mediaInputKind';

export interface MediaInput {
  name: string;
  contentType: string;
  /** @minimum 1 */
  size: number;
  kind: MediaInputKind;
}