# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  slug: string;
  name: string;
  price: string;
  description: string;
  features: string[];
}