# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleInput {
  /** @minLength 1 */
  make: string;
  /** @minLength 1 */
  model: string;
  /**
     * @minimum 1900
     * @maximum 2100
     */
  year: number;
  engine: string;
  /** @minimum 0 */
  mileage: number;
  fuel: string;
  transmission: string;
  registration: string;
  /** @nullable */
  vin?: string | null;
}